moved game logic to game-of-life
