+++ game-of-life-tdd-js/src/game-of-life/cell_state.js
